Migrate EventConfig to AppDbConfig
