+++ quickevent/app/quickevent/plugins/CardReader/src/cardreaderwidget.cpp
@@ -488,7 +488,7 @@ void CardReaderWidget::reset()
 
 void CardReaderWidget::reload()
 {
-	bool is_relays = getPlugin<EventPlugin>()->eventConfig()->isRelays();
+	bool is_relays = getPlugin<EventPlugin>()->appDbConfig()->isRelays();
 	//QString driver_name = m_cardsModel->sqlConnection().driverName();
 	int current_stage = getPlugin<CardReaderPlugin>()->currentStageId();
 	qfs::QueryBuilder qb;
