+++ quickevent/app/quickevent/plugins/Event/src/appdbconfig.h
@@ -1,0 +1,79 @@
+#pragma once
+
+#include "eventconfig.h"
+
+#include <QObject>
+#include <QVariantMap>
+#include <QSet>
+#include <QDateTime>
+
+#include <optional>
+
+namespace Event {
+
+class AppDbConfig
+{
+public:
+	enum class Sport {OB = 1, LOB, MTBO, TRAIL};
+	enum class Discipline {LongDistance = 1,
+						   ShortDistance = 2,
+						   Sprint = 3,
+						   UltralongDistance = 4,
+						   Relays = 5,
+						   Teams = 6,
+						   FreeOrder = 7,
+						   NightRace = 9,
+						   TempO = 11,
+						   MultiStages = 13,
+						   MassStart = 14,
+						   SprintRelays = 15,
+						   KnocOutSprint = 16,
+						   Indoor = 19,
+						  };
+	static std::optional<Discipline> disciplineFromInt(int i);
+public:
+	explicit AppDbConfig() = default;
+public:
+	QVariantMap values() const {return m_data;}
+	QVariant value(const QStringList &path, const QVariant &default_value = QVariant()) const;
+	QVariant value(const QString &path, const QVariant &default_value = QVariant()) const {
+		return value(path.split('.'), default_value);
+	}
+	void setValue(const QStringList &path, const QVariant &val);
+	void setValue(const QString &path, const QVariant &val) {setValue(path.split('.'), val);}
+	void load();
+	void save(const QString &path_to_save = QString());
+
+	int stageCount() const;
+	int currentStageId() const;
+	int sportId() const;
+	Discipline discipline() const;
+	int importId() const;
+	int handicapLength() const;
+	bool isHandicap() const {return handicapLength() > 0;}
+	bool isRelays() const {
+		return discipline() == Discipline::Relays
+				|| discipline() == Discipline::Teams
+				|| discipline() == Discipline::SprintRelays;
+	}
+	bool isIofRace() const;
+	int iofXmlRaceNumber() const;
+	QString eventName() const;
+	QString eventPlace() const;
+	QString director() const;
+	QString mainReferee() const;
+	QDateTime eventDateTime() const;
+	int dbVersion() const;
+	std::optional<int> maximumCardCheckAdvanceSec() const;
+	bool isOneTenthSecResults() const;
+	QString orisEventKey() const;
+	EventConfig eventConfig() const;
+private:
+	void save_helper(QVariantMap &ret, const QString &current_path, const QVariant &val);
+	QVariantMap setValue_helper(const QVariantMap &m, const QStringList &path, const QVariant &val);
+private:
+	QVariantMap m_data;
+};
+
+}
+
