+++ quickevent/app/quickevent/plugins/Event/src/eventconfig.h
@@ -1,78 +1,39 @@
-#ifndef EVENTCONFIG_H
-#define EVENTCONFIG_H
+#pragma once
 
-#include <QObject>
+#include "stagedata.h"
+
+#include <QDate>
+#include <QMap>
+#include <QString>
+#include <QTime>
 #include <QVariantMap>
-#include <QSet>
-#include <QDateTime>
-
-#include <optional>
 
 namespace Event {
 
-class EventConfig
+struct EventConfig
 {
-public:
-	enum class Sport {OB = 1, LOB, MTBO, TRAIL};
-	enum class Discipline {LongDistance = 1,
-						   ShortDistance = 2,
-						   Sprint = 3,
-						   UltralongDistance = 4,
-						   Relays = 5,
-						   Teams = 6,
-						   FreeOrder = 7,
-						   NightRace = 9,
-						   TempO = 11,
-						   MultiStages = 13,
-						   MassStart = 14,
-						   SprintRelays = 15,
-						   KnocOutSprint = 16,
-						   Indoor = 19,
-						  };
-	static std::optional<Discipline> disciplineFromInt(int i);
-public:
-	explicit EventConfig() = default;
-public:
-	QVariantMap values() const {return m_data;}
-	QVariant value(const QStringList &path, const QVariant &default_value = QVariant()) const;
-	QVariant value(const QString &path, const QVariant &default_value = QVariant()) const {
-		return value(path.split('.'), default_value);
-	}
-	void setValue(const QStringList &path, const QVariant &val);
-	void setValue(const QString &path, const QVariant &val) {setValue(path.split('.'), val);}
-	void load();
-	void save(const QString &path_to_save = QString());
+	static EventConfig fromVariantMap(const QVariantMap &values);
+	QVariantMap toVariantMap() const;
 
-	int stageCount() const;
-	int currentStageId() const;
-	int sportId() const;
-	Discipline discipline() const;
-	int importId() const;
-	int handicapLength() const;
-	bool isHandicap() const {return handicapLength() > 0;}
-	bool isRelays() const {
-		return discipline() == Discipline::Relays
-				|| discipline() == Discipline::Teams
-				|| discipline() == Discipline::SprintRelays;
-	}
-	bool isIofRace() const;
-	int iofXmlRaceNumber() const;
-	QString eventName() const;
-	QString eventPlace() const;
-	QString director() const;
-	QString mainReferee() const;
-	QDateTime eventDateTime() const;
-	int dbVersion() const;
-	std::optional<int> maximumCardCheckAdvanceSec() const;
-	bool isOneTenthSecResults() const;
-	QString orisEventKey() const;
-private:
-	void save_helper(QVariantMap &ret, const QString &current_path, const QVariant &val);
-	QVariantMap setValue_helper(const QVariantMap &m, const QStringList &path, const QVariant &val);
-private:
-	QVariantMap m_data;
+	int stageCount = 1;
+	QMap<int, StageData> stages;
+	QString name;
+	QDate date;
+	QTime time;
+	QString description;
+	QString place;
+	QString mainReferee;
+	QString director;
+	int handicapLength = 0;
+	int sportId = 0;
+	int disciplineId = 0;
+	int importId = 0;
+	QString orisEventKey;
+	int cardCheckTimeSec = 0;
+	int oneTenthSecResults = 0;
+	bool iofRace = false;
+	int iofXmlRaceNumber = 0;
+	int currentStageId = 1;
 };
 
 }
-
-#endif // EVENTCONFIG_H
