+++ quickevent/app/quickevent/plugins/Event/src/stagedata.h
