+++ quickevent/app/quickevent/plugins/Runs/src/runstablemodel.cpp
@@ -62,7 +62,7 @@ Qt::ItemFlags RunsTableModel::flags(const QModelIndex &index) const
 		//qfInfo() << flgs;
 	}
 	if(index.column() == col_course_id) {
-		if (getPlugin<Event::EventPlugin>()->eventConfig()->isRelays()) {
+		if (getPlugin<Event::EventPlugin>()->appDbConfig()->isRelays()) {
 			flgs &= ~Qt::ItemIsEditable;
 		}
 	}
@@ -72,7 +72,7 @@ Qt::ItemFlags RunsTableModel::flags(const QModelIndex &index) const
 QVariant RunsTableModel::data(const QModelIndex &index, int role) const
 {
 	if(index.column() == col_course_id && role == Qt::DisplayRole) {
-		if (getPlugin<Event::EventPlugin>()->eventConfig()->isRelays()) {
+		if (getPlugin<Event::EventPlugin>()->appDbConfig()->isRelays()) {
 			auto start_number = value(index.row(), "startNumber").toInt();
 			auto leg = value(index.row(), "runs.leg").toInt();
 			return QStringLiteral("%1.%2").arg(start_number).arg(leg);
